% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zi_kw_test}
\alias{zi_kw_test}
\title{Modified Kruskal Wallis test for zero-inflated data}
\usage{
zi_kw_test(x, g, alpha = 0.05, perm = FALSE)
}
\arguments{
\item{x}{a numeric vector of data values, or a list of numeric data vectors.}

\item{g}{a vector or factor object giving the group for the corresponding elements of x. Ignored if x is a list.}

\item{alpha}{the significant level}

\item{perm}{use permutation}
}
\value{
pvalue
}
\description{
Modified Kruskal Wallis test for zero-inflated data
}
\examples{
zi_kw_test(x, g, alpha = 0.05, perm = FALSE)
}

