% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zi_wilcox_test}
\alias{zi_wilcox_test}
\title{Modified Wilcoxon rank test for zero-inflated data}
\usage{
zi_wilcox_test(x, y, alpha = 0.05, perm = FALSE)
}
\arguments{
\item{x}{the covariates for zero proportion}

\item{y}{the covariates for non-zero compoment}

\item{alpha}{the significant level}

\item{perm}{use permutation}
}
\value{
pvalue
}
\description{
Modified Wilcoxon rank test for zero-inflated data
}
\examples{
zi_wilcox_test(x, y, alpha = 0.05, perm = FALSE)
}

