% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_kw_test.R
\name{zikw}
\alias{zikw}
\title{Modified Kruskal Wallis test (ZIKW) for zero-inflated data}
\usage{
zikw(x, group, perm = FALSE, perm.n = 10000, seed = NULL)
}
\arguments{
\item{x}{a numeric vector of data values with groups defined by the group parameter, or a list of numeric data vectors with each vector in the list as one group.}

\item{group}{a vector giving the groups for the corresponding elements of \code{x}. Ignored if \code{x} is a list.}

\item{perm}{use permutations to calculate the p-value}

\item{perm.n}{number of permutations used for p-value calculation}

\item{seed}{integer to set the random seed (see also \code{\link{set.seed}}); default is NULL (no set seed)}
}
\value{
modified Kruskal Wallis test statistic and p-value
}
\description{
Modified Kruskal Wallis test (ZIKW) for zero-inflated data
}
\examples{
## x is a list
x <- list(group1 = c(rep(0, 5),rlnorm(20, meanlog = 0, sdlog = 1)),
     group2=c(rep(0, 10),rlnorm(20, meanlog = 1, sdlog = 1)),
     group3=c(rep(0, 15),rlnorm(20, meanlog = 2, sdlog = 1)))
zikw(x, perm = FALSE)
## x is a vector
x <- c(c(rep(0, 5),rlnorm(20, meanlog = 0, sdlog = 1)),
    c(rep(0, 10),rlnorm(20, meanlog = 1, sdlog = 1)),
    c(rep(0, 15),rlnorm(20, meanlog = 2, sdlog = 1)))
group <- c(rep('group1', 25),rep('group2', 30),rep('group3', 35))
zikw(x, group, perm = FALSE)
## use permutations to calculate the p-value
\dontrun{zikw(x, group, perm = TRUE)}
}
\references{
Wanjie Wang, Eric Z. Chen and Hongzhe Li (2021). Rank-based tests for compositional distributions with a clump of zeros. Submitted.
}
