% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_wilcox_test.R
\name{ziw}
\alias{ziw}
\title{Modified Wilcoxon rank test (ZIW) for zero-inflated data}
\usage{
ziw(x, y, perm = FALSE, perm.n = 10000, seed = NULL)
}
\arguments{
\item{x}{a vector of data for one group}

\item{y}{a vector of data for another group}

\item{perm}{logical indicating use of permutation to calculate the p-value}

\item{perm.n}{number (integer) of permutations used for p-value calculation}

\item{seed}{integer to set the random seed (see also \code{\link{set.seed}}); default is NULL (no set seed)}
}
\value{
modified wilcoxon rank sum statistic and p-value
}
\description{
Modified Wilcoxon rank test (ZIW) for zero-inflated data
}
\examples{
x <- c(rep(0, 5), rlnorm(20, meanlog = 0, sdlog = 1))
y <- c(rep(0, 10), rlnorm(20, meanlog = 2, sdlog = 1))
ziw(x, y, perm = FALSE)
## use permutations to calculate the pvalue
ziw(x, y, perm = TRUE)
}
\references{
Wanjie Wang, Eric Z. Chen and Hongzhe Li (2021). Rank-based tests for compositional distributions with a clump of zeros. Submitted.
}
