% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zi_kw_test.R
\name{zikw}
\alias{zikw}
\title{Modified Kruskal Wallis test (ZIKW) for zero-inflated data}
\usage{
zikw(x, group, perm = FALSE, perm.n = 10000)
}
\arguments{
\item{x}{a numeric vector of data values with groups defined by the group parameter, or a list of numeric data vectors with each vector in the list as one group.}

\item{group}{a vector giving the groups for the corresponding elements of x. Ignored if x is a list.}

\item{perm}{use permutations to calculate the pvalue}

\item{perm.n}{number of permutations used for pvalue calculation}
}
\value{
modified Kruskal Wallis test statistic and pvalue
}
\description{
Modified Kruskal Wallis test (ZIKW) for zero-inflated data
}
\examples{
## x is a list
x <- list(group1 = c(rep(0,5),rlnorm(20, meanlog = 0, sdlog = 1)),
     group2=c(rep(0,10),rlnorm(20, meanlog = 1, sdlog = 1)),
     group3=c(rep(0,15),rlnorm(20, meanlog = 2, sdlog = 1)))
zikw(x, perm = FALSE)
## x is a vector
x <- c(c(rep(0,5),rlnorm(20, meanlog = 0, sdlog = 1)),
    c(rep(0,10),rlnorm(20, meanlog = 1, sdlog = 1)),
    c(rep(0,15),rlnorm(20, meanlog = 2, sdlog = 1)))
group <- c(rep('group1',25),rep('group2',30),rep('group3',35))
zikw(x, group, perm = FALSE)
## use permutations to calculate the pvalue
\dontrun{zikw(x, group, perm = TRUE)}
}

