% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zi_wilcox_test.R
\name{ziw}
\alias{ziw}
\title{Modified Wilcoxon rank test (ZIW) for zero-inflated data}
\usage{
ziw(x, y, perm = FALSE, perm.n = 10000)
}
\arguments{
\item{x}{a vector of data for one group}

\item{y}{a vector of data for another group}

\item{perm}{use permutation to calculate the pvalue}

\item{perm.n}{number of permutations used for pvalue calculation}
}
\value{
modified wilcoxon rank sum statistic and pvalue
}
\description{
Modified Wilcoxon rank test (ZIW) for zero-inflated data
}
\examples{
x <- c(rep(0,5),rlnorm(20, meanlog = 0, sdlog = 1))
y <- c(rep(0,10),rlnorm(20, meanlog = 2, sdlog = 1))
ziw(x, y, perm = FALSE)
## use permutations to calculate the pvalue
ziw(x, y, perm = TRUE)
}
\references{
Wanjie Wang, Eric Z. Chen and Hongzhe Li (2015). Rank-based tests for compositional distributions with a clump of zeros. Submitted.
}

